Fix: Resolve all 73 clippy warnings across codebase
This commit addresses all clippy warnings to achieve a clean build:

- vibesql-types: Remove empty lines after doc comments
- vibesql-catalog: Add allow attribute for too_many_arguments
- vibesql-parser: Add allow attributes for type_complexity
- vibesql-storage: Fix file operations, doc formatting, use clamp(),
  replace map().flatten() with and_then(), add allow for method names
- vibesql-executor: Remove duplicate cfg attributes, use strip_prefix(),
  add allow attributes for collapsible_match and type_complexity
- vibesql-cli: Auto-fixed by clippy

All changes maintain functionality while improving code quality.

+++ crates/vibesql-storage/src/page/mod.rs
@@ -158,7 +158,7 @@ impl PageManager {
     pub fn read_page(&self, page_id: PageId) -> Result<Page, StorageError> {
         let mut file = lock!(self.file);
 
-        let offset = (page_id as u64) * (PAGE_SIZE as u64);
+        let offset = page_id * (PAGE_SIZE as u64);
         let mut data = vec![0u8; PAGE_SIZE];
 
         match file.read_at(offset, &mut data) {
@@ -178,7 +178,7 @@ impl PageManager {
     pub fn write_page(&self, page: &mut Page) -> Result<(), StorageError> {
         let mut file = lock!(self.file);
 
-        let offset = (page.id as u64) * (PAGE_SIZE as u64);
+        let offset = page.id * (PAGE_SIZE as u64);
         file.write_at(offset, &page.data)?;
         file.sync_data()?;
